# backend/api/openapi/openapi-product.yaml
openapi: 3.1.0
info:
  title: Product Configuration API
  version: "0.1.0-draft"
  description: >
    External REST surface of the `product` module. Product authoring (create/publish
    versions, rating tables) is a staff/actuarial back-office capability; the
    read/catalog side is what customer and agent channels actually consume
    (product browsing, quote calculation inputs).
servers:
  - url: https://api.nlolo-lifeplatform.tz/v1
paths:
  /products:
    get:
      summary: List active products (catalog browsing)
      security:
        - customersAuth: [profile]
        - agentsAuth: [agent]
        - staffAuth: []
      parameters:
        - name: category
          in: query
          schema: { type: string, enum: [TERM_LIFE, ENDOWMENT, WHOLE_LIFE, ANNUITY, UNIT_LINKED, GROUP_LIFE, EDUCATION_SAVINGS, CREDIT_LIFE] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProductSummary' }
    post:
      summary: Create a new product definition (staff/actuarial only)
      security:
        - staffAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        '201':
          description: Product created (status DRAFT)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductSummary' }
  /products/drafts:
    get:
      summary: List DRAFT products — created but never published (staff/actuarial only)
      description: >-
        Authoring is two phases: create the definition, then publish a version. Only the
        second flips DRAFT to ACTIVE, and `GET /products` returns ACTIVE only — so a product
        whose second phase was abandoned appeared in no list while `ux_product_code` kept
        holding its code. The code was burned and the product could be neither seen nor
        finished. This is the list that makes it reachable, so the publish form can be
        reopened against it.


        A separate path rather than a `status` parameter on `GET /products`, because that
        endpoint answers to customersAuth and agentsAuth: an unlaunched product is not
        something a policyholder or a tied agent gets to enumerate. This one is staff-only
        and, in the implementation, ADMIN-only — the same role that may create a product.
      security:
        - staffAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProductSummary' }
  /products/{productId}:
    get:
      summary: One product by id — its code, name, category and status
      description: >-
        The lookup that turns a product id into a product NAME. Nothing did:
        `/products/{productId}/active-snapshot` takes an id but its view carries pricing and
        eligibility, never the code or the name, and `GET /products` carries both but is
        keyed by nothing. So a caller holding an id had to fetch the whole catalogue and
        scan it, and got nothing at all for a product that is DRAFT or has been retired out
        of the catalogue.


        Every console screen referencing a product it had not itself chosen from a list
        therefore printed the raw uuid: the underwriting queue rendered a column of them,
        and the case detail rail printed one under the label "Product".


        Any status, deliberately — a retired product is the case that needs this most, since
        it is gone from the catalogue while the policies and cases referencing it are still
        on screen. Same security as `GET /products`, and NOT the staff-only gate on
        `/products/drafts`: what that one protects is ENUMERATION of unlaunched products,
        and resolving a single id the caller already holds enumerates nothing.
      security:
        - customersAuth: [profile]
        - agentsAuth: [agent]
        - staffAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductSummary' }
        '404':
          description: "`PRODUCT_NOT_FOUND` -- no such product in this tenant"
          content:
            application/problem+json:
              schema: { $ref: 'openapi-common.yaml#/components/schemas/ProblemDetails' }
  /products/{productId}/versions:
    post:
      summary: Publish a new product version (staff/actuarial only)
      description: Rejected with 422 if ifrsMeasurementModel is not set (Deliverable 3 Rev 2 — carried from Deliverable 2 Rev 2's mandatory-before-ACTIVE rule).
      security:
        - staffAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductVersionSpec' }
      responses:
        '201':
          description: Version published
        '422':
          description: Missing ifrsMeasurementModel, incomplete rating-factor coverage, or a fund definition on a non-unit-linked product
          content:
            application/problem+json:
              schema: { $ref: 'openapi-common.yaml#/components/schemas/ProblemDetails' }
  /products/{productId}/active-snapshot:
    get:
      summary: Get the version active as of a given date (used for quoting and rating)
      description: >-
        A product in the catalogue does NOT imply a version in force today. Publishing a
        version flips the definition to ACTIVE whatever its effective date, and a version
        carries a retirement date -- so a product can be listed while its only version
        starts next week, or after every version it ever had has retired. Both are normal
        states of product authoring, not errors in the request.


        Which is why the two 404s below carry different error codes. They were one code
        (`PRODUCT_NOT_FOUND`) and the version lookup ran before the existence check, so
        "its version starts tomorrow" was reported as "no such product" -- and the console,
        which must hedge on a bare 404 because one can be a disguised denial, told a staff
        member the record did not exist or their role forbade it, immediately after they
        clicked it in the catalogue.
      security:
        - customersAuth: [profile]
        - agentsAuth: [agent]
        - staffAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: effectiveDate
          in: query
          description: Defaults to today. The date the version is resolved as of.
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductSnapshot' }
        '404':
          description: >-
            `PRODUCT_NOT_FOUND` -- no such product. `NO_ACTIVE_PRODUCT_VERSION` -- the
            product exists and no version of it is in force on that date; the detail names
            the date asked about and, where one is scheduled, the date the next version
            takes effect.
          content:
            application/problem+json:
              schema: { $ref: 'openapi-common.yaml#/components/schemas/ProblemDetails' }
  /products/{productId}/premium-quote:
    post:
      summary: Price a product for one applicant, on the version active as of a date
      description: >-
        200, not 201: nothing is created. Persists nothing, publishes nothing, and
        takes NO Idempotency-Key -- it is a calculation, so replaying it is free and
        there is no duplicate to prevent.

        POST rather than GET despite being a read, because the body carries
        dateOfBirth, sex and smokerStatus; a GET would put those in a URL and
        therefore into access logs, proxy logs and browser history.

        Resolution is STRICT: a version with no base rate table, an age outside every
        band, or a declared occupation class / sum-assured band with no multiplier
        each return 422 naming the dimension. Nothing falls back to a neutral value,
        because a fallback here produces a plausible premium for a real contract and
        nothing fails.
      security:
        - agentsAuth: [agent]
        - staffAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PremiumQuoteRequest' }
      responses:
        '200':
          description: Priced
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PremiumQuoteView' }
        '401': { $ref: 'openapi-common.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'openapi-common.yaml#/components/responses/Forbidden' }
        '404': { description: "No product, or no version active as of that date" }
        '422': { $ref: 'openapi-common.yaml#/components/responses/UnprocessableEntity' }
  /products/{productId}/versions/{versionId}/rating:
    get:
      summary: The rating basis for one version, for actuarial and product review
      description: >-
        Deliberately separate from ProductSnapshot, which Underwriting and Billing
        read on the issuance and billing path: folding rate tables into it would make
        every consumer carry data exactly one screen needs, and it could not be
        role-scoped because Billing must keep reading it. This is staff-only.
      security:
        - staffAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: versionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionRatingView' }
        '401': { $ref: 'openapi-common.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'openapi-common.yaml#/components/responses/Forbidden' }
        '404': { description: "No such version for that product in this tenant" }
  /products/{productId}/versions/{versionId}/exclusion-periods:
    put:
      summary: Set or clear a version's policy-term exclusion windows
      description: >-
        How long each exclusion stays open, in months from cover start. Without this operation
        the windows could only be set by writing SQL, which left the claims exclusion gate inert
        in any real deployment: a version ships with both null, no window is ever open, and every
        exclusion decline an assessor tries to record is refused for citing one that is not.

        For credit life that is not a minor gap. Below the free cover limit nobody is
        underwritten -- and with the limit far above any loan these lenders write, nobody ever is
        -- so these two windows are the entire anti-selection control the product has. The client
        confirmed 12 and 12 on 2026-09-22.

        PUT because it is idempotent and replaces both values together. Sending one field alone
        CLEARS the other, deliberately: these are the version's exclusions as a whole, and a
        partial update would make "no suicide exclusion" indistinguishable from "I did not
        mention it". A body of {} is therefore a legitimate request meaning the product excludes
        neither, not a malformed one.

        Read them back through GET /products/{productId}/active-snapshot, which already carries
        both fields.
      security:
        - staffAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: versionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetExclusionPeriodsRequest' }
      responses:
        '204': { description: Set }
        '400': { $ref: 'openapi-common.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'openapi-common.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'openapi-common.yaml#/components/responses/Forbidden' }
        '404': { description: "No such version in this tenant" }
components:
  schemas:
    ProductSummary:
      type: object
      properties:
        productId: { type: string, format: uuid }
        productCode: { type: string }
        productName: { type: string }
        category: { type: string }
        status: { type: string, enum: [DRAFT, ACTIVE, RETIRED] }
        defaultCurrency: { type: string, example: TZS }
    CreateProductRequest:
      type: object
      required: [productCode, productName, category, defaultCurrency]
      properties:
        productCode: { type: string }
        productName: { type: string }
        category: { type: string, enum: [TERM_LIFE, ENDOWMENT, WHOLE_LIFE, ANNUITY, UNIT_LINKED, GROUP_LIFE, EDUCATION_SAVINGS, CREDIT_LIFE] }
        defaultCurrency: { type: string, example: TZS }
    EligibilityBounds:
      description: >-
        What this version will accept. Every bound optional -- an unbounded dimension is a real
        product design, not a gap.

        The bounds are NOT all the same kind of rule, and the gates that consume them differ.
        Entry age is a HARD refusal: base_rate_table is keyed on (age range, sex, smoker status)
        and a quote already 422s with PremiumNotQuotableException when no cell matches, so an
        out-of-range age cannot be priced at all. Term is a HARD refusal too: term is
        staff-entered configuration, and a term the product was not designed for is the wrong
        product. Sum assured is a SOFT flag: above retention is not invalid business, it is what
        reinsurance is for, and the breach is recorded in reasonForManualIssue rather than
        blocked.

        Sum-assured bounds carry no currency -- they are read in the product's own
        defaultCurrency, and a bound in a currency the product does not price in means nothing.
      type: [object, "null"]
      properties:
        minEntryAge: { type: [integer, "null"], minimum: 0, maximum: 120 }
        maxEntryAge: { type: [integer, "null"], minimum: 0, maximum: 120 }
        minTermMonths: { type: [integer, "null"], minimum: 1 }
        maxTermMonths: { type: [integer, "null"], minimum: 1 }
        minSumAssured: { type: [number, "null"], exclusiveMinimum: 0 }
        maxSumAssured: { type: [number, "null"], exclusiveMinimum: 0 }
...
    PremiumQuoteRequest:
      description: >-
        Takes dateOfBirth, never a precomputed age: entry age is the input a
        mispriced policy turns on, so it is derived where the rate table lives.
        `occupationClass` and `smokerStatus` are ASSERTED by the caller -- no party
        record carries either.

        There is deliberately no `sumAssuredBand`: it was asserted by the caller and
        matched by string equality against rating_table.band, which is the defect V9
        removed from the issuance path. `sumAssuredAmount` determines the band, and
        the multiplier is resolved from it by range.
      type: object
      required: [sumAssuredAmount, sumAssuredCurrency, dateOfBirth, sex, smokerStatus, occupationClass, frequency]
      properties:
        sumAssuredAmount: { type: number, exclusiveMinimum: 0 }
        sumAssuredCurrency: { type: string, example: TZS }
        dateOfBirth: { type: string, format: date }
        sex: { type: string, enum: [FEMALE, MALE] }
        smokerStatus: { type: string, enum: [SMOKER, NON_SMOKER, UNKNOWN] }
        occupationClass: { type: string }
        # SINGLE carries no frequency loading -- a loading prices the cost of spreading
        # payment across the year, and a single premium spreads nothing. Its instalment is
        # the whole loaded annual figure, charged once, so `instalmentsPerYear` comes back 0
        # rather than 1: there is no recurring period, and a 1 would read as "annually" to
        # any arithmetic that divides by it.
        #
        # The annual figure IS the single premium only while the cover term is twelve months.
        # A single premium over a longer policyTermMonths is refused (422) rather than quoted at
        # the wrong number, until a single-premium basis is added.
        frequency: { type: string, enum: [MONTHLY, QUARTERLY, ANNUALLY, SINGLE] }
        asOf:
          description: Selects the product version. Defaults to today.
          type: [string, "null"]
          format: date
        policyTermMonths:
          description: >-
            The policy term the quote is for (V16). Optional: null prices against unbanded rates
            only and is refused on a term-banded version, the same rule issuance follows.
          type: integer
          minimum: 1
    PremiumQuoteView:
      description: >-
        The premium AND its derivation. The breakdown is returned because
        "reproduce exactly what the customer was shown" is a requirement, and
        without it a UI could only show one by recomputing it. No policy fee is
        included: no product field holds one.
      type: object
      properties:
        productVersionId: { type: string, format: uuid }
        currency: { type: string, example: TZS }
        ageAtEntry: { type: integer }
        ageFrom: { type: integer }
        ageTo: { type: integer }
        ratePerMille: { type: number }
        annualBase: { type: number }
        appliedFactors:
          type: array
          items:
            type: object
            properties:
              factorType: { type: string, enum: [AGE, OCCUPATION_CLASS, SMOKER_STATUS, SUM_ASSURED_BAND] }
              band: { type: string }
              multiplier: { type: number }
        annualAfterFactors: { type: number }
        frequencyLoadingPercent: { type: number }
        annualAfterFrequencyLoading: { type: number }
        # 0 for SINGLE, and the instalment beside it is then the whole premium. See the
        # request schema's own note.
        frequency: { type: string, enum: [MONTHLY, QUARTERLY, ANNUALLY, SINGLE] }
        instalmentsPerYear: { type: integer }
        instalmentAmount: { type: number }
...
    SetExclusionPeriodsRequest:
      type: object
      additionalProperties: false
      description: >-
        Both fields nullable, and null clears that window -- a product with no such exclusion is
        the normal case and has to stay expressible, which is why neither is required. The 120
        upper bound is not arbitrary: a window longer than a decade is not a policy term, it is a
        typo, and the cost of accepting one is a claim declined years after any assessor would
        defend the decision.
      properties:
        suicideExclusionMonths: { type: [integer, "null"], minimum: 0, maximum: 120 }
        preExistingExclusionMonths: { type: [integer, "null"], minimum: 0, maximum: 120 }
    ProductSnapshot:
      type: object
      properties:
        productId: { type: string, format: uuid }
        productVersionId: { type: string, format: uuid }
        effectiveDate: { type: string, format: date }
        ifrsMeasurementModel: { type: string, enum: [GMM, PAA] }
        gracePeriodDays: { type: integer }
        maxLoanToValuePercent: { type: number }
        category: { type: string, enum: [TERM_LIFE, ENDOWMENT, WHOLE_LIFE, ANNUITY, UNIT_LINKED, GROUP_LIFE, EDUCATION_SAVINGS, CREDIT_LIFE] }
        surrenderChargeScheduleJson: { type: [string, "null"], description: Raw JSON surrender-charge schedule from product_version.surrender_charge_schedule; null until a future milestone adds a way to author it. }
        suicideExclusionMonths:
          type: [integer, "null"]
          description: >-
            Months from cover start during which an assessor may cite suicide as a decline reason.
            Null where the product has no such exclusion. An OPEN window is permission to cite a
            reason, never a bar on settlement -- an ordinary death inside one is paid in full,
            because this product has no general waiting period.
        preExistingExclusionMonths:
          type: [integer, "null"]
          description: >-
            The same, for a pre-existing condition. Null where the product has no such exclusion.
        eligibility:
          description: >-
            What this version will accept. Always present; an unbounded version is the object with
            six nulls. Read by the issue screen's gates, where entry age and term are hard refusals
            and sum assured is a soft flag.
          allOf: [{ $ref: '#/components/schemas/EligibilityBounds' }]